+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## Unreleased
 - Fixed - Use of std::wstring_convert class (deprecated C++17, removed C\++26)
-- Added - cvt.toString(...) support for wchar_t and wstring
+- Added - cvt.toString() support for wchar_t and wstring
+- Added - cli.toArgv() and cli.toCmdline() variants that report errors on
+          cvt.toString() failures.
 
 ## dimcli 7.5.0 (2025-11-20)
 - Added - cli.success() and cli.fail() overloads with just the exit code